server/requirements.txt
# FastAPI + server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1

# Auth & security (bcrypt directly to avoid passlib/bcrypt version incompatibility)
bcrypt>=4.0,<5
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
email-validator==2.2.0

# Optional: real email when SMTP_* env vars are set (stdlib smtplib used)
